Provide a "vz venv" command which calls the actual virtualenv
This is done in order to make it easier choosing a virtualenv
implementation which runs on Python 3.x to be used for "vz new",
but can be used manually as well.

This solves issues with the virtualenv command not being found
in systems where the program names are always versioned, e.g.
in OpenBSD 6.x